chore: migrate Rstest and Rspress to Rstack CLI
