+++ package.json
@@ -57,7 +57,6 @@
     "pnpm": "11.8.0"
   },
   "devDependencies": {
-    "@rstest/core": "catalog:",
     "@taplo/cli": "catalog:",
     "@types/node": "catalog:",
     "@typescript/native": "catalog:",
