--- packages/rspack-cli/rstest.config.ts
@@ -1,34 +1,0 @@
-import { defineConfig, type RstestConfig } from '@rstest/core';
-
-const wasmConfig: RstestConfig | undefined = process.env.WASM
-  ? {
-      exclude: [
-        '**/*/profile.test.ts', // Skip due to lack of system api support
-      ],
-    }
-  : undefined;
-
-export default defineConfig({
-  name: 'rspack-cli',
-  testEnvironment: 'node',
-  globals: true,
-  testTimeout: process.env.CI ? 200000 : 30000,
-  include: ['tests/**/*.test.{ts,js,cts}'],
-  source: {
-    tsconfigPath: 'tests/tsconfig.json',
-  },
-  output: {
-    externals: [/@rspack\/core/, /api-wrapper/],
-    module: false,
-  },
-  env: {
-    RUST_BACKTRACE: 'full',
-  },
-  chaiConfig: process.env.CI
-    ? {
-        // show all info on CI
-        truncateThreshold: 5000,
-      }
-    : undefined,
-  ...(wasmConfig || {}),
-});
