+++ website/tsconfig.json
@@ -21,5 +21,5 @@
   "mdx": {
     "checkMdx": true
   },
-  "include": ["src", "docs", "theme", "components", "./rspress.config.ts"]
+  "include": ["src", "docs", "theme", "components", "./rstack.config.ts"]
 }
